+++ src/triangle.cpp
@@ -8,8 +8,8 @@
 
 namespace triangles {
 void createTriangles(const flock::Flock& flock, sf::VertexArray& triangles) {
-  for (int i = 0; i < flock.getBoidsNum(); ++i) {
-    const int j = 3 * i;
+  for (size_t i = 0; i < flock.getBoidsNum(); ++i) {
+    const size_t j = 3 * i;
     sf::Vertex vertex{flock.getBoidFlock()[i]->getPosition()()};
 
     // upper vertex (centered)
@@ -26,10 +26,10 @@ void createTriangles(const flock::Flock& flock, sf::VertexArray& triangles) {
   }
 
   if (flock.getPredatorsNum() > 0) {
-    for (int i = 0; i < flock.getPredatorsNum(); ++i) {
+    for (size_t i = 0; i < flock.getPredatorsNum(); ++i) {
       sf::Vertex vertex{flock.getPredatorFlock()[i]->getPosition()()};
 
-      const int j = 3 * (i + flock.getBoidsNum());
+      const size_t j = 3 * (i + flock.getBoidsNum());
       // upper vertex (centered)
       triangles[j].position = vertex.position + sf::Vector2f(0, -(height * 1.5) / 2);
       triangles[j].color = sf::Color::Red;
@@ -43,17 +43,17 @@ void createTriangles(const flock::Flock& flock, sf::VertexArray& triangles) {
       triangles[j + 2].color = sf::Color::Red;
     }
   }
-  assert(static_cast<int>(triangles.getVertexCount()) == flock.getFlockSize() * 3);
+  assert(triangles.getVertexCount() == flock.getFlockSize() * 3);
 }
 
-void rotateTriangle(const point::Point& target_position, sf::VertexArray& triangles, const double theta, const int i,
-                    const bool is_boid, const int nBoids) {
+void rotateTriangle(const point::Point& target_position, sf::VertexArray& triangles, const double theta, const size_t i,
+                    const bool is_boid, const size_t nBoids) {
   const sf::Vertex vertex{target_position()};
 
-  const int j = is_boid ? 3 * i : 3 * (i + nBoids);
+  const size_t  j = is_boid ? 3 * i : 3 * (i + nBoids);
 
-  for (int k = j; k < j + 3; ++k) {
-    const int index = is_boid ? k - j : k - j + 3;
+  for (size_t k = j; k < j + 3; ++k) {
+    const size_t index = is_boid ? k - j : k - j + 3;
     triangles[k].position =
         vertex.position + sf::Vector2f(static_cast<float>(relative_position[index].x * std::cos(theta) -
                                                           relative_position[index].y * std::sin(theta)),
